Create README.md
The Auto Heading Tags Plugin for WordPress was born out of a personal need to streamline content creation for Infocero, a news platform. One of the most important aspects of SEO and content readability is the proper use of heading tags (H1, H2, H3, etc.) to structure your posts. However, manually managing these tags for each post can be time-consuming and prone to errors, especially when dealing with large volumes of content, as is often the case with news websites.